# Various allocation optimizations