These are 3 independent optimizations to avoid various (client-side) allocations I found during profiling.

The first two commits are purely to avoid allocations, the third one optimizes the Metadata.Entry check and normalization by using a simple loop and char comparison instead of the Regex. It also avoids a string allocation if the input is already lowercase (which should be the common case). The benchmark code for this change can be found here: https://gist.github.com/szehetner/c2044f703783bc7a19706cf7ead5805d

Results:
```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.16299.611 (1709/FallCreatorsUpdate/Redstone3)
Intel Xeon CPU E5-1620 v2 3.70GHz, 1 CPU, 8 logical and 4 physical cores
Frequency=3613277 Hz, Resolution=276.7571 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3132.0
  DefaultJob : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3132.0


         Method | InputValue |      Mean |     Error |     StdDev | Scaled |  Gen 0 | Allocated |
--------------- |----------- |----------:|----------:|-----------:|-------:|-------:|----------:|
        ToLower | HeaderName | 474.05 ns | 5.5187 ns |  4.8922 ns |   1.00 | 0.0086 |      48 B |
 ToLowerNoRegex | HeaderName |  94.15 ns | 1.7107 ns |  1.5165 ns |   0.20 | 0.0091 |      48 B |
                |            |           |           |            |        |        |           |
        ToLower | headername | 482.84 ns | 9.4752 ns | 11.6364 ns |   1.00 | 0.0086 |      48 B |
 ToLowerNoRegex | headername |  15.38 ns | 0.2342 ns |  0.2191 ns |   0.03 |      - |       0 B |
```