@grpc-testing 2018-09-28
```
Objective-C binary sizes
*****************STATIC******************
  New size                      Old size
 1,964,802      Total (=)      1,964,802

 No significant differences in binary sizes

***************FRAMEWORKS****************
  New size                      Old size
10,887,474      Total (>)     10,887,470

 No significant differences in binary sizes


```